Move mal_simulator.py to root package directory instead of in ./sims which is now removed
