--- malsim/sims/__init__.py
@@ -1,19 +1,0 @@
-from .mal_simulator import (
-    MalSimulatorSettings,
-    MalSimulator,
-    AgentType,
-    MalSimAgentState,
-    MalSimDefenderState,
-    MalSimAttackerState,
-    MalSimAgentStateView
-)
-
-__all__ = [
-    'AgentType',
-    'MalSimAgentState',
-    'MalSimAgentStateView',
-    'MalSimDefenderState',
-    'MalSimAttackerState',
-    'MalSimulatorSettings',
-    'MalSimulator',
-]
