+++ tests/agents/test_searchers.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 from maltoolbox.attackgraph import AttackGraphNode, Attacker
 from maltoolbox.attackgraph.query import get_attack_surface
 from maltoolbox.language import LanguageGraph
-from malsim.sims import MalSimAgentStateView
+from malsim.mal_simulator import MalSimAgentStateView
 from malsim.agents import BreadthFirstAttacker, DepthFirstAttacker
